requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
scikit-surprise>=1.1.1
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
flask>=2.0.1
gunicorn>=20.1.0
flask-cors>=3.0.10